OMAPDSS: SiI9022: Add magic sleep to fix EDID read
We often see failed EDID reads, especially the first one done at boot
time. The root cause is unknown, but empiric testing shows that a small
sleep between the read of the first EDID block and the read of the
second EDID block makes the problem go away.

This patch adds a 10ms sleep, and a comment explaining the issue.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>