ARM: dts: OMAP4: Add L2RAM info to IPU node
The L2RAM memory region information has been added to the
IPU node through the 'reg' and 'reg-names' properties. This
change also requires the corresponding OF_DEV_AUXDATA entry
to be updated in pdata-quirks to be able to match the updated
node. The target device name is unchanged though to not require
any changes in the driver.

Note that the DSP node is not updated since it doesn't have
any L2 RAM memory, but only Unicaches.

Signed-off-by: Suman Anna <s-anna@ti.com>