remoteproc/pruss: add pruss_get()/put() API
Add two new get and put API, pruss_get() and pruss_put(), to the
PRUSS platform driver to allow client drivers to request a handle
to a PRUSS device. This handle will be used by client drivers to
request various operations of the PRUSS platform driver through
additional API that will be added in the following patches.

The pruss_get() function returns the pruss handle corresponding
to a PRUSS device referenced by the PRUSS phandle in the "pruss"
property of the client device's DT node. The pruss_put() is the
complimentary function to pruss_get().

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>
[s-anna@ti.com: error checking, few other minor fixes]
Signed-off-by: Suman Anna <s-anna@ti.com>