drm/omap: omap_wb_m2m: Fix an instance count error
In omap_wb_m2m when device open fails the atomic instance counter
was not decrement leaving it as if an instance was already in use.
Also since the required dss plane is allocated on the first open
only it left the reserved omap_plane_id as invalid causing kernel oops
on subsequent attempt to open the device.

Also in the failure case make sure that the m2m context gets
cleaned up properly.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>