drm/omap: omap_wb_m2m: Fix an infinite loop when streaming does not start
commit 0f1d2bb97e91 branch ti-linux-4.1.y

In cases where an error is cwreported/discovered from inside
device_run() and seince device_run() should not fail the
m2m context is left in the pending state with nothing clean
it up, in this case v4l2_m2m_cancel_job() would wait forever
for the "job" to finish.

Hence we are forcing the job to finish in abort cases.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>