drm/omap: omap_wb_m2m: Fix a frame size buffer alignment issue
commit 81a8b61cd165 branch ti-linux-4.1.y

The buffer alignment check was over zealously requiring 16 bytes
alignment and also requiring buffer height to be even.
These requirements were too stringent for this module.
In this instance it is sufficient that the line stride be even
with the exception of YUV 422 interleaved format which require
an even number of pixels per line.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>