pinctrl: single: Change alloc in save_context to GFP_ATOMIC
pcs_save_context, introduced in commit a46f3359aecd ("ARM: OMAP2:
Add functions to save and restore pinctrl context.") must be called
late in the suspend sequence during modes when the hardware will lose
context because many drivers may modify the pinctrl values earlier in
the sequence. Because of this, it gets called with irqs disabled so we
must use GFP_ATOMIC when doing the initial allocation for space to save
the existing pinctrl values, otherwise we will see a sleeping from
invalid context bug when lock debugging is enabled.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>