remoteproc/pruss: drop pru_trigger_interrupt() function
The function pru_trigger_interrupt() function is defined to
raise an event to trigger an interrupt on one of the PRUSS
host interrupts, but is currently not used anywhere. If it
ends up being needed in the future, the function can always
be added back. Until then, let's drop unused code.

Signed-off-by: Felipe Balbi <balbi@ti.com>
[afd@ti.com: rebase function]
Signed-off-by: Andrew F. Davis <afd@ti.com>
[s-anna@ti.com: revise commit description]
Signed-off-by: Suman Anna <s-anna@ti.com>