soc: ti: Allow ti_sci_pm_domains driver to be built as a module
The ti_sci_pm_domains driver depends on TI_SCI_PROTOCOL which may
be built as a module, so this also needs to be able to be built as
a module.

Change TI_SCI_PM_DOMAINS to tristate and export an OF MODULE_ALIAS.
This is only loaded through OF so the platform MODULE_ALIAS is not
needed.

Signed-off-by: Andrew F. Davis <afd@ti.com>