HACK: remoteproc/pru: do not auto-boot PRU cores on IDKs
The PRUSS Ethernet usecase is currently designed to have the ethernet
driver perform the boot of a PRU device (in the expected fashion when
a firmware doesn't support virtio devices). The commit ad6bf515e79f
("remoteproc/pruss: manage PRUs without virtio devices manually") has
provided a temporary mechanism to manage the boot cycle of PRU cores
running firmwares without any virtio devices previously for generic
usecases.

It cannot be currently determined whether the PRUs are being used for
an Ethernet usecase or the previously supported generic usecases, so
make a decision using the assumption that Ethernet usecases are to
be supported only on IDK boards. This assumption itself is not solid
though and is not scalable, it is only made on the premises that the
supported IDKs will be running only Ethernet usecases.

Real fix will need restructuring work done on the remoteproc core and
a rework on the PRUSS remoteproc driver architecture to meet its
varying usecases.

Signed-off-by: Andrew F. Davis <afd@ti.com>
[s-anna@ti.com: revise commit description]
Signed-off-by: Suman Anna <s-anna@ti.com>