net: netcp: extract eflag from desc for rx_hook handling
As a preparatory patch, extract the eflag bits from the received
desc and pass it down the rx_hook chain to be available for
netcp modules. Also the psdata and epib data has to be inspected
by the netcp modules. So the desc can be freed only after
returning from the rx_hook. So move knav_pool_desc_put() after
the rx_hook processing.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>