ARM: dts: AM33xx/AM437x/DRA7: Rename PRU nodes
Each of the PRU child nodes has been renamed from the
generic name 'pru' to a core-specific 'pru0' or 'pru1'
name. This allows them to be identified and looked up
specifically using name.

NOTE:
The DT specification in general does not require that
different instances be named accordingly, but this is
being done at present to support the new API that will
be added to support the PRUSS Ethernet usecases. The
normal way to use a device would be to use the phandle
approach in DT.

Signed-off-by: Andrew F. Davis <afd@ti.com>
[s-anna@ti.com: revise binding and commit description]
Signed-off-by: Suman Anna <s-anna@ti.com>