ARM: dts: am572x-idk: Add mailboxes to IPU and DSP nodes
Add the required mailboxes to all the remote processor device
nodes (IPU1, IPU2, DSP1 and DSP2) for the AM572x IDK board. This
is required for running the Remote Processor Messaging (RPMsg)
stack between the host processor and each of the remote processors.
This data is identical to that of the equivalent processors on the
AM57xx EVM board to maintain firmware compatibility between both
the SoC and board variants.

Each of the remote processors uses a single sub-mailbox node, the
IPUs are assumed to be running in SMP-mode. The chosen sub-mailboxes
match the values used in the current firmware images. This can be
changed, if needed, as per the system integration needs after making
appropriate changes on the firmware side as well.

Signed-off-by: Suman Anna <s-anna@ti.com>