ARM: dts: am571x-idk: Add watchdog timers for IPU2
The watchdog timer information has been added to the IPU2 remote
processor device node on AM571x IDK board. This data is identical
to the timers used on the DRA72 EVM board to maintain firmware
compatibility between both the SoC and board variants.

The watchdog functionality will use GPTimers 4 and 9 for each of
the cores in IPU2. No watchdog timer is added to the remaining
IPU and DSP processors, but can be easily added as per need.

The MPU-side drivers will use this data to initialize the watchdog
timers, and listen for any watchdog triggers. The BIOS-side code on
IPU2 needs to configure/refresh this timer properly to not throw a
watchdog error.

Signed-off-by: Suman Anna <s-anna@ti.com>