OMAPDSS: make panels depend on OMAP_DSS_BASE
omapdss-base module was created some time ago, which contains the
support for panel drivers to register themselves. omapdss-base is
enabled when OMAP_DSS_BASE Kconfig symbol is set.

All the panel and encoder drivers still depend on OMAP2_DSS, which is
not correct, and prevents from using the drivers with the upcoming DSS6
driver. So fix the Kconfig dependency.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>