arm: dts: k2g-evm.dts: add data for HDMI
Add DT data for K2G EVM's HDMI display.

The HDMI setup on K2G EVM is identical to HDMI setup on TI AM4 EVMs: DSS
parallel RGB output goes to Silicon Image's SiI9022 HDMI encoder, which
goes to a HDMI connector.

There's also a TPD12S016 ESD protection chip in between the SiI9022 and
the HDMI connector, but as the CT_HPD and LS_OE pins on the chip have
been hardwired to high, and thus there's no way to control the device,
it's not represented in the DT data.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>