net: netcp: pa: add timestamp support
The PDSP firmware available on PA hardware is capable of doing timestamp
on tx/rx packets. It implements the PAHO_PAMOD_REPORT_TIMESTAMP command
to do this. For tx packets, the driver add this command to each packet
handed over to the PA h/w and driver process the responses that has the
timestamp information. On the receive side, each packet will have the
timestamp counter which is retrieved and reported by the driver.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>