ARM: DTS: k2g-evm: Enable HDMI audio support
The HDMI audio setup is: McASP2 -> si9022. The protocol is i2s on the
digital audio bus.

NOTE: the same McASP is also used for analog audio (with aic3106 codec).
HDMI and analog audio can not be used at the same time! We have the very
same McASP2 used for analog and HDMI audio...

If audio started while the other audio card is playing audio you will
hear the already playing audio through the card you started the new
playback (since ASoC will configure the codec), but DMA will not start
as it is already running and we will have timeout on the new stream.
This is going to stop the new stream and going to shut down the McASP2
as well, so the older playback will break also.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>