ARM: dts: dra7-evm: add support for higher MMC/SD speed
Silicon revision 2.0 of DRA74x SoC, available on revision
H and onwards of DRA74x EVM provides for higher MMC/SD
speeds.

This is possible because errata i843 is addressed on this
silicon.

Update device-tree data to utilize these higher speeds.
IOdelay data for both silicon revision 1.1 and 2.0 is
maintained. The maximum frequency supported is updated to
match that of rev 2.0. pdata quirks must be utilized to
detect presence of rev 1.1/1.0 of silicon and adjust
maximum frequencies as per restrictions documented in i843.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>