ARM: OMAP: DRA7: powerdomain data: Remove CSWR with the exception of MPU
As per the latest definitions for DRA7/AM57xx SoCs, the internal
memories of SoC do not support retention mode any longer (with the
exception of MPU), as a result of the same, we also end up losing CSWR
(where powerdomain state can be programmed to retention state).

This implies that, with the exception of MPU power domain, all other
power domains can either operate in "ON" mode OR where possible "OFF"
mode (when permitted by lack of silicon errata).

Signed-off-by: Nishanth Menon <nm@ti.com>