HACK: remoteproc/pru: use different firmware names for IDKs
The PRUSS Ethernet usecase is currently designed to have the PRUSS
ethernet driver perform the load and boot of a PRU device (in the
expected fashion when a firmware doesn't support virtio devices),
though the ethernet driver wants to use its own firmware name. The
remoteproc core instrastructure currently does the loading as a
two-step process with the resource table processed and cached the
first step, and a CRC-based cross-check in the second step. The
ethernet driver fails to load an image properly due to either a
mismatched resource table or a complete absence of a resource table.

Add a work-around to the PRU remoteproc driver to enable the Ethernet
usecases by using matching firmware names for supported IDKs. This
is based on the assumption that the supported IDKs will be running
only Ethernet usecases for the moment. This solution was preferred
rather than making any remoteproc core changes so as to not affect
other existing remoteproc drivers.

Real fix will need restructuring work done on the remoteproc core to
facilitate client provided firmware names at run-time and not rely
on firmware names provided at respective platform driver probe time.

Signed-off-by: Suman Anna <s-anna@ti.com>