net/rpmsg: fix a use after free error
The rpmsg_proto probe function assigns a freed sock_list
pointer variable after it has been freed when the radix
tree insertion fails. Fix this by properly bailing out
on an error. Defect is discovered using the Coverity code
coverage tool.

Note that the patch is not fixing any real issue as the
driver core will unwind whatever is done on the device,
and the rpmsg-proto driver will never use this rpmsg
device or dereference the assigned pointer.

Signed-off-by: Karthik Ramanan <a0393906@ti.com>
[s-anna@ti.com: revise commit description]
Signed-off-by: Suman Anna <s-anna@ti.com>