ARM: dts: am572x-idk: add support for UHS cards
Rev 1.3A of the AM572x IDK supports UHS mode cards.
This is made possible by the use of revision 2.0 of
AM572x silicon and fixes on the board allowing
voltage switching to happen.

MMC on previous versions of the board are not
guaranteed to work after this change since the IODelay
values added here are for revision 2.0 of AM572x
silicon. But those are unsupported anyway and
should all be replaced by these newer board versions.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>