media: ti-vpe: cal: Restructure where the camera core mapping is done
Since we now need to be able to touch the Camera Core registers
of either phy regardless of the fact that there is a sensor
connected or not, then we need to relocate the camera core register
mapping area to the main device structure instead of being only
available on a per context basis.

The context structure now only has a back pointer to the appropriate
instance in the main device struct.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>