remoteproc/pruss: remove dead rproc_put() in pruss probe
The PRUSS platform driver doesn't have anything to do with the
management of a rproc device, that is the responsibility of the
PRU remoteproc platform driver. However, the pruss probe code has
a stale dead code fragment of rproc_put(). This should have never
been present to begin with, so clean it up.

Signed-off-by: Suman Anna <s-anna@ti.com>