ARM: OMAP2+: Add am33xx_pm_deinit to disable cpuidle
Currently if pm33xx is built as a module and removed, cpuidle does not
get disabled so if it attempts to enter C1 state, the call into pm33xx
will result in a hang. Add hooks to allow unregister of cpuidle33xx
driver so that we can remove it when pm33xx module gets removed.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>