dmaengine: virt-dma: Support for ignoring callback after vc termination
This patch provides means to prevent the race condition between vchan
completion tasklet and the terminate_all function.
The race condition is not easily reproducible, but it has been observed
with RT kernel on uniprocessor systems with high rate DMA activity (UART)
and it follows this pattern:
When terminate_all is called, the tasklet has been already scheduled or
it is running to handle the completion of a vdesc. The client driver might
be freeing up resources after terminating the transfers but the tasklet
when it has a chance to continue to execute will call the provided callback
even after the client driver does not expect this to happen.
The terminated flag for the vchan itself does not provide enough protection
since can still face with a race condition:
in this case when the vchan_complete() tests for the terminated flag it will
see that the vchan is not terminated, but right after the test it is
possible that the terminate thread is set to execute (uniprocessor + RT) and
it will free up the resources. After that the vchan_complete() will resume
and calls the callback. This will lead to crash.

To provide protection, the dma driver need to make sure that the tasklet
will not run while we terminate the channel by:

static int dmadriver_terminate_all(struct dma_chan *chan)
{
	unsigned long flags;
	LIST_HEAD(head);
...
	tasklet_disable(&vc.task);
	spin_lock_irqsave(&vc.lock, flags);
	vchan_terminate(&vc);
...
	vchan_get_all_descriptors(&vc, &head);
	spin_unlock_irqrestore(&vc.lock, flags);
	vchan_dma_desc_free_list(&vc, &head);
	tasklet_enable(&vc.task);

	return 0;
}

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>