media: ti-vpe: vip: Split parser access to its own memory region
As of now the VIP driver maps the its entire memory space in to
a single region as only the VPDMA section it stand alone.
In preparation to add scaler and color space support for which
stand alone memory region are mandated by the common library code
we are then splitting every sub-module into its own memory region
starting with the parser.

The existing code was therefore reworked a little to take advantage
of this split.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>