media: ti-vpe: vip: Add scaler support
Adding support for the scaler unit.
Since there is only one scaler unit shared between 2 ports
a design decision was made that only one active port would have
access to the scaler unit. The other port would behave as usual
relying of the sensor capability to provide different frame sizes.

When a device node is opened  if the scaler unit is unassigned
(i.e. available) then is assigned to that port only if that port is
not configured for multiple streams already.

Once the scaler unit is made available subsequent try/set format
will assume scaling is part of the data path.

As before given the output format the frame size is constraint
to vpdma alignment restriction. (i.e. bytes per line must be 16
bytes aligned).

As usual, v4l2 application using this device must make sure to
pay attention to the returned width, height, bytesperline and
sizeimage buffer parameters to behave properly.

For example a 600x400 in NV12 format will yield a 608x400 frame
size instead because 600 is not divisible by 16 the closest value
is 608.

As we have discovered up-scaling vertically causes serious artifact
to the output stream because the scaling buffer cannot stall the
incoming stream.

The frame size selection make sure that the selected sub device
frame size is at least as large as the requested size,
but not larger.
The sub device frame size selection is always performed
even when scaler is enabled in order to find a supported
frame size that can be scale down from.

Trying to request a frame size that would result in vertical
scaling will result in selecting the largest sub device frame
size available instead.

Signed-off-by: Benoit Parrot <bparrot@ti.com>