ARM: DRA7: hwmod: Make rng hwmod non-secure only
If running on a non-secure (GP) device, the hwmod for the
hardware RNG should be initialized. It should NOT be used
on secure (HS) devices as it will be firewalled for secure
world use. The RNG is not as nice as other crypto modules (AES,
SHA) which have secure and public interfaces, with the public
interfaces opened to the public world (Linux kernel) and the
secure interfaces firewalled to be used by the secure world only.

Giving access to the RNG from the public world could pose a
security risk for the secure world (by reducing the randomness
available through manipulation of the RNG config), therefore
the firewall is set to enforce no public access on HS devices.

Signed-off-by: Daniel Allred <d-allred@ti.com>