remoteproc/pruss: fix up the INTC configuration per PRU core
The PRUSS INTC configuration is currently stored globally in the PRUSS
structure, and is filled in when each PRU core is loaded through the
processing of its resource table entry. This configuration is never
undone when a PRU core is unloaded.

Fix up the configuration of the PRUSS INTC to be managed per PRU core.
This is done by adding a pruss_intc_unconfigure() function and reworking
the current pruss_intc_configure() function. These functions will only
be used by the pru_rproc platform driver during the booting/shutdown of
the PRU. So, the per PRU core INTC configuration is stored in the
pru_rproc structure, and passed in to the INTC configuration functions.

This fixes the loading and unloading of the pru_rproc module or the
unloading and reloading of a specific PRU core. This allows different
applications to unload and reload a PRU.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>