remoteproc/pru: change firmware names for PRUSS Ethernet
Change the firmware names to be used for the PRUSS Ethernet EMAC
functionality on AM437x and AM57xx IDK boards to reflect the
location of the firmwares on the filesystem and to match their
path on the TI linux-firmware tree.

The ti-prueth driver also needs to match these names.

Fixes: a7868b69a2e3 ("HACK: remoteproc/pru: use different firmware names for IDKs")
Signed-off-by: Suman Anna <s-anna@ti.com>