net: prueth: change firmware names for PRUSS Ethernet
The PRUSS Ethernet driver would require different firmwares to
be loaded on different SoCs. Update the ti-prueth driver to use
different compatibles based on the SoC families, so that the
appropriate firmwares can be loaded on the different IDK boards.
Updates have been made to handle AM437x, AM571x and AM572x SoCs,
with AM571x and AM572x using the same firmwares.

he existing DTS node for AM57xx IDKs has been updated accordingly
as well. The firmware names are also updated to match the location
of the firmwares on the filesystem and to match their path on the
TI linux-firmware tree.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>