rpmsg: rpc: define and use a device address type
The current rpmsg_rpc code uses a phys_addr_t type for variables
storing the remote processor device addresses. The phys_addr_t
type can either be a 32-bit or a 64-bit integer type (based on
CONFIG_PHYS_ADDR_T_64BIT). Define a new address type, dev_addr_t,
and use this for remote processor addresses instead of phys_addr_t,
so as to not to confuse with either the actual physical addresses
or the MPU virtual addresses. The new type is defined to be a 32-bit
integer type as all the existing OMAP remote processors use 32-bit
addresses only.

This fixes couple of build warnings of the type "cast to pointer from
integer of different size" when CONFIG_ARM_LPAE is enabled.

Signed-off-by: Suman Anna <s-anna@ti.com>