net/rpmsg: fix autoloading of rpmsg-proto driver
The rpmsg-proto is a rpmsg bus driver that provides support
for a new remote-processor socket protocol, and exposes a
rpmsg communication channel to userspace through the socket
API. It resides in the networking layer, but is not a true
networking driver.

The driver is not probed based on devices from DT, but rather
the devices are published to the kernel when a firmware is
loaded by a remoteproc driver. The existing MODULE_ALIAS_NETPROTO
doesn't trigger an autoload of the driver by udev. So, add a
rpmsg-specific MODULE_ALIAS that fixes the autoloading of the
rpmsg-proto driver.

Signed-off-by: Suman Anna <s-anna@ti.com>