rpmsg: rpc: fix return values in rppc_local_to_remote_da
The function rppc_local_to_remote_da() is used for translating
kernel local buffer addresses to remote processor device addresses,
and is expected to return a value of 0 on errors. This translation
is done using the rproc_da_to_va() exported function from the
remoteproc core, with a rproc handle looked up using the current
rpc instance. However, it is returning an uninitialized value at
present when the rproc handle lookup fails, or -EINTR if the mutex
locking is interrupted. This results in a kernel crash further
along due to operation on an invalid buffer address value.

Fix this by returning a value of 0 on all possible error paths.
Also, updated the comments with details about the return values.

Reported-by: Buddy Liong <a0270631@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>