ARM: dts: dra7: do not gate cpsw clock due to errata i877
Do not allow to gate the cpsw clocks using ti,no-idle property in
cpsw node as a workaround for errata i877.

Due to this errata, the device switching characteristics
degrade over time when cpsw clock is gated and eventually
failing to meet the data manual delay time/skew specifications.

The prescribed workaround for this is for the software to
minimize the time Ethernet internal clock source is kept
disabled.

Signed-off-by: Mugunthan V N <mugunthanvnm@ti.com>
[grygorii.strashko@ti.com: fwp from TI K4.1]
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>