dmaengine: omap-dma: Fix completion reporting via tx_status
When based on the CCR_ENABLE bit the channel is stopped we should not call
omap_dma_callback(), only change the return value to DMA_COMPLETE. Client
drivers will do the right thing to clean up the channel after the transfer
has been completed.
Check the CCR_ENABLE only if the channel is running and it is not paused:
If the channel is not yet started the CCR_ENABLE bit is not set.
sDMA can not be paused. Pause (in cyclic mode only) means that we stop
the channel.

This will fix one hard to reproduce race condition when the channel is
terminated during transfer (affecting cyclic operation).

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>