remoteproc/pruss: fix MStandby configuration for system suspend
The PRU-ICSS subsystem has a separate PRUSS_CFG module that
contains various configuration registers. This includes a control
bit STANDBY_INIT in register PRUSS_CFG register to initiate a
Standby sequence (when set) and trigger a MStandby request to the
SoC's PRCM module. This same bit is also used to enable the OCP
master ports (when cleared). The system suspend/resume functionality
on AM33xx/AM437x SoCs requires all initiators to assert their
MStandby signal properly inorder to properly enter suspend, and
resume on a wakeup event.

Commit 8f54b75363d9 ("remoteproc/pruss: enable OCP master ports in
SYSCFG") unconditionally enables the OCP master ports to enable the
PRUSS Ethernet usecase. This caused a hang in the resume sequence
on AM33xx/AM437x boards and required a board reset to come out of
the hang.

This patch adds the preliminary System PM callbacks, and fixes this
system resume hang by setting the STANDBY_INIT in the PM system
suspend callback and resetting it in the PM system resume callback.
The logic to set the STANDBY_INIT and checking for the subsequence
SUB_MWAIT ready signal is refactored into a common function so that
both the pruss_resume and pruss_enable_module functions can reuse.

NOTE:
1. This patch only adds the PM callbacks with code to fix the System
   Suspend/Resume hang issue on AM33xx/AM437x SoCs, but does not
   implement the full context save and restore required for the PRUSS
   driver to work across system suspend/resume when the power domain
   is switched off (L4PER domain is switched OFF on AM335x/AM437x
   during system suspend/resume, so PRUSS modules do lose context).
2. The PRUSS driver functionality on AM57xx SoCs is not affected
   because the PER power domain to which the PRUSS IPs belong is not
   switched OFF during suspend/resume.

Fixes: 8f54b75363d9 ("remoteproc/pruss: enable OCP master ports in SYSCFG")
Signed-off-by: Suman Anna <s-anna@ti.com>