spi: spi-ti-qspi: Add DMA support for QSPI mmap read
Since QSPI support mmap-read, mem-to-mem DMA can be used to increase
throughput and decrease CPU load. But, the JFFS2 FS doesn't provide
DMA'able buffers, in which case pre-allocated bounce buffer is used.

With this patch the raw-read throughput is ~16MB/s on DRA74 EVM. And CPU
load is <20%. UBIFS read ~13 MB/s.

Signed-off-by: Vignesh R <vigneshr@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>