usb: dwc3: core: Introduce dwc3_device_reinit()
commit 7e96adaba3c0d4b4abd00857275a66efa737fb78 ti-linux-4.1.y

We will need this function for a workaround.
The function issues a softreset only to the device
controller and performs minimal re-initialization
so that the device controller can be usable.

As some code is similar to dwc3_core_init() take out
common code into dwc3_device_soft_reset() and
dwc3_get_gctl_quirks().

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>

Conflicts:
	drivers/usb/dwc3/core.c