net: netcp: fix the Ethernet receive issue on 2U switch
For the newer version of the switch such as NU and 2U, map
all of the vlan priorities to zero as the packet DMA is enabled
to receive only flow id 0 which maps to priority zero. Without
this, the switch gets into a rx fifo overrun at the slave interface.
This patch fixes this issue.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>