ARM: dts: K2G: Use the correct DCAN clock
The clock index for DCAN is currently set to 0 which references
the DCAN's VBUS clk which runs at 100 MHz. However, what we really
want is the DCAN's CAN clk which runs at 24 MHz.

Correct this so that the DCAN driver gets the proper clock rate. Also
instead of just giving a number use the proper define so its clear exactly
what clock we wanted to use.

Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>
Acked-by: Roger Quadros <rogerq@ti.com>