soc: ti: enhancement to support qmss on nss lite
nss Lite available on K2G SoC has a cut down version of qmss.
It doesn't have status and explicit push register space as in
qmss available on other K2 SoCs. So define reg indices specific
to qmss on nss lite (call it qmss lite). This patch introduces
"keystone-navigator-qmss-l" compatibility to identify this
version of the qmss and to customize the dts handling code.

Update the Documentation for DT bindings to reflect this change

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Acked-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>