ARM: dts: dra7xx-clocks: Set DSP DPLL and its output clock rates
The DSP DPLL is a new DPLL compared to previous OMAP generations and
is currently not configured by bootloader. This DPLL supplies the root
clocks for the DSP processors, as well as a mux input source for EVE
sub-system (on applicable SoCs). Use the DT standard properties
"assigned-clocks" and "assigned-clock-rates" to set the DSP DPLL clock
rate and the rates for its derivative clocks at boot time to avoid any
functional issues caused when using any of the DSP processors if the
DPLL is not locked. The DPLL will automatically transition into a
low-power stop mode when the associated output clocks are not utilized
or gated automatically.

The DSP DPLL provides two output clocks, DSP_GFCLK and EVE_GCLK. The
desired rate for DSP_GFCLK is 600 MHz (same as DSP DPLL CLKOUT frequency),
and is currently auto set due to the desired M2 divider value being the
same as reset value for the locked frequency of 600 MHz. The EVE_GCLK
however is required to be 400 MHz, so set the dpll_dsp_m3x2_ck's rate
explicitly so that the divider is set properly. The dpll_dsp_m2_ck rate
is also set explicitly to not rely on any implicit matching divider reset
values to the locked DPLL frequency.

The clock rates are chosen based on the OPP_NOM values defined as per the
DRA7xx PLL spec v0.4WIPa. The DPLL locked frequency is 1200 MHz, so the
dpll_dsp_ck clock rate used is half of this value.

Signed-off-by: Suman Anna <s-anna@ti.com>