ARM: dts: am572x-idk: Add pcie-reset-gpio property to control PCIE_RESETn
On AM572x IDK, GPIO3_23 is connected to the GPIO_PCIE_SWRSTn which in
turn is connected to PCIE_PERSTn through a non inverting buffer.

Add 'pcie-reset-gpio' property to pcie1 dt node and populate it with
GPIO3_23 in order to drive PCIE_RESETn high.

This gets PCIe cards to be detected in AM572X IDK board.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>