Revert "media: ti-vpe: vip: Add scaler support"
This reverts commit b4397f835dc4ded28c2fa7bfd4194c5ae78d96e9.

Although the VIP scaler supports upscaling, any work being done
by the scaler is on the incoming line buffer. This means that
if the scaler is still processing by the time the next line
comes in you end up with distortion/shift of various severity
depending on the scaling factor in the vertical direction.

Scaling up horizontally is somewhat safer, but when the scaler
has to "generate" several lines for each incoming input lines
the data will most likely get overwritten while the scaler is
still busy with it.

The current patch does not prevent scaling up. Some extra logic
is needed to prevent these now known non working cases.

In the mean time it is safer and quicker to disable the
feature for now until we can solidify the use cases.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>